Reorganized project package/file structure. Renamed ProjectManager -> ServerManager.
