+++ src/main/java/com/revature/schoolDatabase/datasource/models/Faculty.java
@@ -1,4 +1,4 @@
-package com.revature.schoolDatabase.models;
+package com.revature.schoolDatabase.datasource.models;
 
 public class Faculty extends Person {
     private String department;
