+++ src/main/java/com/revature/schoolDatabase/datasource/repositories/CourseRepository.java
@@ -1,12 +1,12 @@
-package com.revature.schoolDatabase.repositories;
+package com.revature.schoolDatabase.datasource.repositories;
 
 import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mongodb.client.*;
 import com.mongodb.client.result.DeleteResult;
 import com.mongodb.client.result.UpdateResult;
-import com.revature.schoolDatabase.models.Course;
-import com.revature.schoolDatabase.util.MongoClientFactory;
+import com.revature.schoolDatabase.datasource.models.Course;
+import com.revature.schoolDatabase.datasource.util.MongoClientFactory;
 import com.revature.schoolDatabase.util.exceptions.DataSourceException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
